[feat] 인증 트랜잭션 범위 축소 - 프록시 내부 호출 관련 작업
