+++ src/main/java/com/nanal/backend/domain/auth/service/AuthService.java
@@ -45,25 +45,25 @@ import static org.springframework.security.oauth2.jwt.JoseHeaderNames.KID;
 @Transactional
 @Service
 public class AuthService {
-    private final MemberRepository memberRepository;
+
     private final TokenUtil tokenUtil;
     private final ClientNaver clientNaver;
     private final ClientKakao clientKakao;
     private final ClientGoogle clientGoogle;
-    private final ApplicationEventPublisher publisher;
     private final AppleFeignClient appleFeignClient;
+
+    private final InternalAuthService internalAuthService;
 
     @Value("${app-id.apple}")
     private String apple_aud;
 
     @Counted("auth.api.count")
-    @DistributedLock
-    public LoginInfo commonAuth(String accessToken, @LockName String providerInfo) {
+    public LoginInfo commonAuth(String accessToken, String providerInfo) {
         // 플랫폼에서 사용자 정보 조회
         Member member = getUserDataFromPlatform(accessToken, providerInfo);
 
         // 회원가입(가입 정보 없는 유저일 때만) 및 로그인
-        Member authenticatedMember = auth(member, providerInfo);
+        Member authenticatedMember = internalAuthService.auth(member, providerInfo);
 
         AuthenticationUtil.makeAuthentication(authenticatedMember);
 
@@ -85,7 +85,7 @@ public class AuthService {
         Member member = Member.createAppleMember(socialId, email);
 
         // 회원가입(가입 정보 없는 유저일 때만) 및 로그인
-        Member authenticatedMember = auth(member, providerInfo);
+        Member authenticatedMember = internalAuthService.auth(member, providerInfo);
 
         AuthenticationUtil.makeAuthentication(authenticatedMember);
 
@@ -116,39 +116,6 @@ public class AuthService {
         if(providerInfo.contains("google")) return clientGoogle.getUserData(accessToken);
         else if(providerInfo.contains("kakao")) return clientKakao.getUserData(accessToken);
         else return clientNaver.getUserData(accessToken);
-    }
-
-    private synchronized Member auth(Member member, String providerInfo) {
-        Optional<Member> findMember = memberRepository.findByEmail(member.getEmail());
-        if(isNewMember(findMember))
-            return joinMembership(member);
-        else
-            return login(findMember, providerInfo);
-    }
-
-    private Member joinMembership(Member member) {
-        Member newMember = register(member);
-        publisher.publishEvent(new RegisterEvent(newMember.getNickname(), newMember.getEmail()));
-        return newMember;
-    }
-
-    private static boolean isNewMember(Optional<Member> findMember) {
-        return findMember.isEmpty();
-    }
-
-    private Member register(Member member) {
-        return memberRepository.save(member);
-    }
-
-    private Member login(Optional<Member> member, String providerInfo) {
-        Member loginMember = member.get();
-        if(!providerInfo.contains((loginMember.getProvider().name().toLowerCase())))
-            throw AccountAlreadyExistException.EXCEPTION;
-
-        if(loginMember.getAlarm() == null) {
-            loginMember.setAlarm(Alarm.createAlarm(loginMember));
-        }
-        return loginMember;
     }
 
     private Jws<Claims> sigVerificationAndGetJws(String unverifiedToken) {
