+++ src/main/java/com/nanal/backend/global/lock/DistributedLockAspect.java
@@ -11,8 +11,6 @@ import org.redisson.api.RedissonClient;
 import org.springframework.core.annotation.Order;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.transaction.support.TransactionSynchronizationManager;
 
 import java.lang.annotation.Annotation;
 import java.time.Duration;
@@ -29,10 +27,11 @@ public class DistributedLockAspect {
     @Around("@annotation(com.nanal.backend.global.lock.DistributedLock)")
     public Object handleLettuceDistributedLock(ProceedingJoinPoint joinPoint) throws Throwable {
         String lockName = getLockName(joinPoint);
+        System.out.println(lockName);
 
+        // try 문 안으로 들어가면 예외 발생시키는 스레드가 임의로 락 해제 시킴
+        if(!lock(lockName)) throw new InternalServerErrorException("lettuce 락 획득 실패");
         try {
-            if(!lock(lockName)) throw new InternalServerErrorException("lettuce 락 획득 실패");
-
             return joinPoint.proceed();
         } catch (InterruptedException e) {
             throw e;
